pyproject.toml
[project]
name = "finance-dashboard"
version = "0.1.0"
description = ""
authors = [
    {name = "Pulkit Govil",email = "pulkitgovil2003@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.135.3,<0.136.0)",
    "motor[srv] (>=3.7.1,<4.0.0)",
    "pymongo (>=4.16.0,<5.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "uvicorn (>=0.43.0,<0.44.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)"
]

[tool.poetry]
packages = [{include = "finance_dashboard", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
